refactor: 모델을 domains 디렉터리로 분리
