+++ __tests__/Lotto.test.js
@@ -1,4 +1,4 @@
-import Lotto from "../src/Lotto";
+import Lotto from "../src/domains/Lotto";
 
 describe("로또 클래스 테스트", () => {
   describe("정상 케이스", () => {
