+++ src/domains/WinningResult.js
